Sean's 9/20 work
implemented a concerning amount of things...

-Moved all driver .c and .h files to their own folders for easier
management
EXAMPLE: ADC_init.c and ADC_init.h are now both in the Drivers/ADC
folder

-Created getAltitude and getVelocity .c and .h files, as well as
methods.

-Created flightState, altitude, pressure, temperature, initialAltitude
variables.

-Cleaned main.c, added #include for new and improved drivers.

-Added switch case for flightState to main.c

-Created flightState class; this houses flightState 0-3 methods, created
a shell for each along with a comment or rough code for what each needs
to do. Also began writing the code for the descent flight state (2).

-Probably also other minor things that I forgot.